Adding SCons build scripts
This patch adds an SCons build system that can build and install CUDD
in a standard location.  It builds a shared and static version of the
library, and installs the necessary header files in the include
directory.  The installation prefix can be specified as follows:

    $ scons prefix=/usr/local
    $ sudo scons install

The first command builds the source and sets the installation prefix
variable; the second actually installs everything.  The build scripts
also check whether the current system is 32-bit or 64-bit, and sets
the CPP macros that CUDD expects.