Using integer as value type for ADDs
This patch modifies the ADD code to use an integer as the value type,
rather than a double.  This is not an ideal fix, but time is of the
essence.  A proper fix would be to create a new set of “IDD” functions
that have integral values, and leave the ADD code to use doubles.